Update directory.html
Created a new branch because I hated the name of the old one.

Edit 1: 
What I did in class...
Still need to figure out how to add elements to a table dynamically. 

Basics of arrays: 
https://www.w3schools.com/js/js_arrays.asp

Example of dynamic table:
https://www.w3schools.com/jsref/tryit.asp?filename=tryjsref_td_create

Specific to what we want to do:
http://stackoverflow.com/questions/18846253/how-do-i-create-html-table-using-jquery-dynamically
http://stackoverflow.com/questions/8302166/dynamic-creation-of-table-with-dom
http://stackoverflow.com/questions/1413952/how-to-create-a-new-table-with-rows-using-jquery-and-wrap-it-inside-div

Edit 2: 

So I added the .append() function to try to add rows to the headers of the tables, but it didn't work. 

JSFiddle version: https://jsfiddle.net/fj57py5z/
JSFiddle version of code that actually works: http://jsfiddle.net/uX6Zq/21/

help me